Creating TabContentScriptManager & TabWebView files
