+++ firefox-ios/Client/TabManagement/Tab.swift
@@ -84,7 +84,13 @@ enum TabUrlType: String {
 typealias TabUUID = String
 
 @MainActor
-class Tab: NSObject, ThemeApplicable, FeatureFlaggable, ShareTab {
+class Tab: NSObject,
+           ThemeApplicable,
+           FeatureFlaggable,
+           ShareTab,
+           ContentBlockerTab,
+           TabWebViewDelegate,
+           UIGestureRecognizerDelegate {
     private var _isPrivate = false
     private(set) var isPrivate: Bool {
         get {
@@ -972,9 +978,9 @@ class Tab: NSObject, ThemeApplicable, FeatureFlaggable, ShareTab {
     func restoreTemporaryDocumentSession(_ session: TemporaryDocumentSession) {
         temporaryDocumentsSession = session
     }
-}
 
-extension Tab: UIGestureRecognizerDelegate {
+    // MARK: - UIGestureRecognizerDelegate
+
     // This prevents the recognition of one gesture recognizer from blocking another
     func gestureRecognizer(
         _ gestureRecognizer: UIGestureRecognizer,
@@ -1006,9 +1012,9 @@ extension Tab: UIGestureRecognizerDelegate {
             )
         }
     }
-}
 
-extension Tab: TabWebViewDelegate {
+    // MARK: - TabWebViewDelegate
+
     func tabWebView(_ tabWebView: TabWebView, didSelectFindInPageForSelection selection: String) {
         tabDelegate?.tab(self, didSelectFindInPageForSelection: selection)
     }
@@ -1024,9 +1030,9 @@ extension Tab: TabWebViewDelegate {
         // Hide the default WKWebView accessory view panel for PDF documents.
         return mimeType != MIMEType.PDF
     }
-}
 
-extension Tab: ContentBlockerTab {
+    // MARK: - ContentBlockerTab
+
     func currentURL() -> URL? {
         return url
     }
@@ -1037,258 +1043,5 @@ extension Tab: ContentBlockerTab {
 
     func imageContentBlockingEnabled() -> Bool {
         return noImageMode
-    }
-}
-
-private class TabContentScriptManager: NSObject, WKScriptMessageHandler {
-    private var helpers = [String: TabContentScript]()
-
-    // Without calling this, the TabContentScriptManager will leak.
-    func uninstall(tab: Tab) {
-        helpers.forEach { helper in
-            helper.value.scriptMessageHandlerNames()?.forEach { name in
-                tab.webView?.configuration.userContentController.removeScriptMessageHandler(forName: name)
-            }
-            helper.value.prepareForDeinit()
-        }
-    }
-
-    func userContentController(_ userContentController: WKUserContentController, didReceive message: WKScriptMessage) {
-        for helper in helpers.values {
-            if let scriptMessageHandlerNames = helper.scriptMessageHandlerNames(),
-               scriptMessageHandlerNames.contains(message.name) {
-                helper.userContentController(userContentController, didReceiveScriptMessage: message)
-                return
-            }
-        }
-    }
-
-    func addContentScript(_ helper: TabContentScript, name: String, forTab tab: Tab) {
-        // If a helper script already exists on a tab, skip adding this duplicate.
-        guard helpers[name] == nil else { return }
-
-        helpers[name] = helper
-
-        // If this helper handles script messages, then get the handlers names and register them. The Browser
-        // receives all messages and then dispatches them to the right TabHelper.
-        helper.scriptMessageHandlerNames()?.forEach { scriptMessageHandlerName in
-            tab.webView?.configuration.userContentController.addInDefaultContentWorld(
-                scriptMessageHandler: self,
-                name: scriptMessageHandlerName
-            )
-        }
-    }
-
-    func addContentScriptToPage(_ helper: TabContentScript, name: String, forTab tab: Tab) {
-        // If a helper script already exists on the page, skip adding this duplicate.
-        guard helpers[name] == nil else { return }
-
-        helpers[name] = helper
-
-        // If this helper handles script messages, then get the handlers names and register them. The Browser
-        // receives all messages and then dispatches them to the right TabHelper.
-        helper.scriptMessageHandlerNames()?.forEach { scriptMessageHandlerName in
-            tab.webView?.configuration.userContentController.addInPageContentWorld(
-                scriptMessageHandler: self,
-                name: scriptMessageHandlerName
-            )
-        }
-    }
-
-    func addContentScriptToCustomWorld(_ helper: TabContentScript, name: String, forTab tab: Tab) {
-        // If a helper script already exists on the page, skip adding this duplicate.
-        guard helpers[name] == nil else { return }
-
-        helpers[name] = helper
-
-        // If this helper handles script messages, then get the handlers names and register them. The Browser
-        // receives all messages and then dispatches them to the right TabHelper.
-        helper.scriptMessageHandlerNames()?.forEach { scriptMessageHandlerName in
-            tab.webView?.configuration.userContentController.addInCustomContentWorld(
-                scriptMessageHandler: self,
-                name: scriptMessageHandlerName
-            )
-        }
-    }
-
-    func getContentScript(_ name: String) -> TabContentScript? {
-        return helpers[name]
-    }
-}
-
-protocol TabWebViewDelegate: AnyObject {
-    @MainActor
-    func tabWebView(_ tabWebView: TabWebView, didSelectFindInPageForSelection selection: String)
-    @MainActor
-    func tabWebViewSearchWithFirefox(
-        _ tabWebViewSearchWithFirefox: TabWebView,
-        didSelectSearchWithFirefoxForSelection selection: String
-    )
-    @MainActor
-    func tabWebViewShouldShowAccessoryView(_ tabWebView: TabWebView) -> Bool
-}
-
-class TabWebView: WKWebView, MenuHelperWebViewInterface, ThemeApplicable, FeatureFlaggable {
-    lazy var accessoryView: AccessoryViewProvider = .build(nil, {
-        AccessoryViewProvider(windowUUID: self.windowUUID)
-    })
-    private var logger: Logger = DefaultLogger.shared
-    private weak var delegate: TabWebViewDelegate?
-    let windowUUID: WindowUUID
-    private var pullRefresh: PullRefreshView?
-    private var theme: Theme?
-
-    override var hasOnlySecureContent: Bool {
-        // TODO: - FXIOS-11721 Understand how it should be showed the lock icon for a local PDF
-        // When PDF is shown we display the online URL for a local PDF so secure content should be true
-        if let url, url.isFileURL, url.lastPathComponent.hasSuffix(".pdf") {
-            return true
-        }
-        return super.hasOnlySecureContent
-    }
-
-    override var inputAccessoryView: UIView? {
-        guard delegate?.tabWebViewShouldShowAccessoryView(self) ?? true else { return nil }
-
-        return accessoryView
-    }
-
-    func configure(delegate: TabWebViewDelegate,
-                   navigationDelegate: WKNavigationDelegate?) {
-        self.delegate = delegate
-        self.navigationDelegate = navigationDelegate
-
-        accessoryView.previousClosure = { [weak self] in
-            guard let self else { return }
-            FormAutofillHelper.focusPreviousInputField(tabWebView: self,
-                                                       logger: self.logger)
-        }
-
-        accessoryView.nextClosure = { [weak self] in
-            guard let self else { return }
-            FormAutofillHelper.focusNextInputField(tabWebView: self,
-                                                   logger: self.logger)
-        }
-
-        accessoryView.doneClosure = { [weak self] in
-            guard let self else { return }
-            FormAutofillHelper.blurActiveElement(tabWebView: self, logger: self.logger)
-            self.endEditing(true)
-        }
-    }
-
-    init(frame: CGRect, configuration: WKWebViewConfiguration, windowUUID: WindowUUID) {
-        self.windowUUID = windowUUID
-        super.init(frame: frame, configuration: configuration)
-    }
-
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    func removeAllUserScripts() {
-        configuration.userContentController.removeAllUserScripts()
-        configuration.userContentController.removeAllScriptMessageHandlers()
-    }
-
-    func menuHelperFindInPage() {
-        ensureMainThread {
-            self.evaluateJavascriptInDefaultContentWorld("getSelection().toString()") { result, _ in
-                let selection = result as? String ?? ""
-                self.delegate?.tabWebView(self, didSelectFindInPageForSelection: selection)
-            }
-        }
-    }
-
-    func menuHelperSearchWith() {
-        ensureMainThread {
-            self.evaluateJavascriptInDefaultContentWorld("getSelection().toString()") { result, _ in
-                let selection = result as? String ?? ""
-                self.delegate?.tabWebViewSearchWithFirefox(self, didSelectSearchWithFirefoxForSelection: selection)
-            }
-        }
-    }
-
-    override internal func hitTest(_ point: CGPoint, with event: UIEvent?) -> UIView? {
-        // The find-in-page selection menu only appears if the webview is the first responder.
-        // Do not becomeFirstResponder on a mouse event.
-        if let event = event, event.allTouches?.contains(where: { $0.type != .indirectPointer }) ?? false {
-            becomeFirstResponder()
-        }
-        return super.hitTest(point, with: event)
-    }
-
-    // swiftlint:disable unneeded_override
-#if compiler(>=6)
-    override func evaluateJavaScript(
-        _ javaScriptString: String,
-        completionHandler: (
-            @MainActor (Any?, (any Error)?) -> Void
-        )? = nil
-    ) {
-        super.evaluateJavaScript(javaScriptString, completionHandler: completionHandler)
-    }
-#else
-    /// Override evaluateJavascript - should not be called directly on TabWebViews any longer
-    /// We should only be calling evaluateJavascriptInDefaultContentWorld in the future
-    @available(*,
-                unavailable,
-                message: "Do not call evaluateJavaScript directly on TabWebViews, should only be called on super class")
-    override func evaluateJavaScript(_ javaScriptString: String, completionHandler: ((Any?, Error?) -> Void)? = nil) {
-        super.evaluateJavaScript(javaScriptString, completionHandler: completionHandler)
-    }
-#endif
-    // swiftlint:enable unneeded_override
-
-    // MARK: - PullRefresh
-
-    func addPullRefresh(onReload: @escaping () -> Void) {
-        guard !scrollView.isZooming else { return }
-        guard pullRefresh == nil else {
-            pullRefresh?.startObservingContentScroll()
-            return
-        }
-        let refresh = PullRefreshView(parentScrollView: scrollView,
-                                      isPortraitOrientation: UIWindow.isPortrait) {
-            onReload()
-        }
-        scrollView.addSubview(refresh)
-        refresh.translatesAutoresizingMaskIntoConstraints = false
-        NSLayoutConstraint.activate([
-            refresh.leadingAnchor.constraint(equalTo: leadingAnchor),
-            refresh.trailingAnchor.constraint(equalTo: trailingAnchor),
-            refresh.bottomAnchor.constraint(equalTo: scrollView.topAnchor),
-            refresh.heightAnchor.constraint(equalTo: scrollView.heightAnchor),
-            refresh.widthAnchor.constraint(equalTo: scrollView.widthAnchor)
-        ])
-
-        refresh.startObservingContentScroll()
-        pullRefresh = refresh
-        guard let theme else { return }
-        refresh.applyTheme(theme: theme)
-    }
-
-    func removePullRefresh() {
-        pullRefresh?.stopObservingContentScroll()
-        pullRefresh?.removeFromSuperview()
-        pullRefresh = nil
-    }
-
-    func setPullRefreshVisibility(isVisible: Bool) {
-        pullRefresh?.isHidden = !isVisible
-    }
-
-    // MARK: - ThemeApplicable
-
-    /// Updates the `background-color` of the webview to match
-    /// the theme if the webview is showing "about:blank" (nil).
-    func applyTheme(theme: Theme) {
-        self.theme = theme
-        backgroundColor = theme.colors.layer1
-        pullRefresh?.applyTheme(theme: theme)
-        if url == nil {
-            let backgroundColor = theme.colors.layer1.hexString
-            evaluateJavascriptInDefaultContentWorld("document.documentElement.style.backgroundColor = '\(backgroundColor)';")
-        }
     }
 }
